requirements.txt
django==4.1.7
djangorestframework==3.14.0
Faker==17.6.0
django-allauth==0.53.0
django-extra-views==0.14.0
